# src/main/resources/static/swagger-ui/openapi3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GOMO API",
    "description": "GOMO API 문서 입니다.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/quests/assigns/calendars": {
      "get": {
        "tags": [
          "Quest"
        ],
        "summary": "할당 퀘스트 캘린더 조회 API",
        "description": "사용자가 한 달간 수행한, 혹은 수행 중인 퀘스트 목록을 조회합니다.",
        "operationId": "calendar_assign_quest",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "description": "년",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "query",
            "description": "월",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarListAssignQuestResponse"
                },
                "examples": {
                  "calendar_assign_quest": {
                    "value": "{\"assignQuests\":[{\"id\":\"0194cbd7-8689-74ec-bd46-dc855f493c3b\",\"questType\":\"DAILY\",\"subjectName\":\"Java\",\"content\":\"daily participating quest\",\"proof\":\"no_proof\",\"completedDateTime\":null,\"completed\":false},{\"id\":\"0194cbda-6135-79fc-b659-ebaac3684761\",\"questType\":\"WEEKLY\",\"subjectName\":\"Java\",\"content\":\"weekly participating quest\",\"proof\":\"no_proof\",\"completedDateTime\":null,\"completed\":false},{\"id\":\"0194cbeb-345e-74a6-9199-07bdb402ea36\",\"questType\":\"MONTHLY\",\"subjectName\":\"Java\",\"content\":\"monthly participating quest\",\"proof\":\"no_proof\",\"completedDateTime\":null,\"completed\":false}]}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarListAssignQuestResponse": {
        "title": "CalendarListAssignQuestResponse",
        "required": [
          "assignQuests"
        ],
        "type": "object",
        "properties": {
          "assignQuests": {
            "type": "array",
            "description": "할당 퀘스트 목록",
            "items": {
              "required": [
                "completed",
                "content",
                "id",
                "proof",
                "questType",
                "subjectName"
              ],
              "type": "object",
              "properties": {
                "questType": {
                  "type": "string",
                  "description": "퀘스트 타입"
                },
                "completedDateTime": {
                  "type": "string",
                  "description": "완료일 (없을 경우 null)",
                  "nullable": true
                },
                "completed": {
                  "type": "boolean",
                  "description": "완료 여부"
                },
                "proof": {
                  "type": "string",
                  "description": "퀘스트 증명"
                },
                "id": {
                  "type": "string",
                  "description": "할당 퀘스트 아이디"
                },
                "content": {
                  "type": "string",
                  "description": "퀘스트 내용"
                },
                "subjectName": {
                  "type": "string",
                  "description": "주제(관심사) 이름"
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "Access-Token"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ]
}